% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{generate_single_profile_biomarker_report}
\alias{generate_single_profile_biomarker_report}
\title{Generates the single response profile biomarker report}
\usage{
generate_single_profile_biomarker_report(out_path, title, Y = NULL,
  meta_data = NULL)
}
\arguments{
\item{out_path}{string path to folder. If Y is NULL this folder should contain biomarker
output files rf_table.csv etc... If Y is given the biomarker output files will be written to this
folder}

\item{title}{string title for the report}

\item{Y}{optional n x p numerical matrix of responses to perturbations,
rownames must be Arxspan IDs (missing values are allowed).}

\item{meta_data}{optional a dataframe containing meta data to include in the report}
}
\description{
Generates the single response profile biomarker report
}
