% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{get_biomarkers}
\alias{get_biomarkers}
\title{Gets the biomarkers for each column in a matrix of responses using:
random_forest
lin_associations
discrete_test}
\usage{
get_biomarkers(Y, p_cutoff = 0.1, out_path = NULL)
}
\arguments{
\item{Y}{n x p numerical matrix of responses to perturbations,
rownames must be Arxspan IDs (missing values are allowed).}

\item{p_cutoff}{maximum p-value in output tables (to keep them smaller), defaults to 0.1}

\item{out_path}{if specified, string path to folder to write results to as a .csv}
}
\value{
a list with components
\describe{
  \item{rf_table}{A table with the following columns:
  feature: the column names of X.
  RF.imp.mean: an estimate of the importance of that feature for model accuracy.
  RF.imp.sd: the standard deviation of the importance estimate.
  RF.imp.stability: the proportion of models that used this feature.
  rank: the rank of the feature in terms of importance for the model.
  MSE: the mean-squared error of the model.
  MSE.se: the standard error of the MSE.
  R2: the \eqn{R^2} of the model.
  PearsonScore: the Pearson correlation of predicted and observed responses.
  pert: the column names of Y
  }
  \item{lin_table}{A table with the following columns:
  feat.A : column names of A
  p.val : p-values computed for the null-hypothesis beta = 0
  q.val : q-values computed using Benjamini-Hotchberg method
  n : number of non-missing values for each column of A
  betahat : estimated linear coefficient controlled for W
  sebetahat ; standard error for the estimate of beta
  lfdr and qvalue: local and global fdr values computed using adaptive shrinkage as 
  described in doi:10.1093/biostatistics/kxw041 
  PosteriorMean and PosteriorSD : moderated effect size estimates and corresponding standard deviations, 
  again computed with adaptive shrinkage.
  pert: the column names of Y
  }
  \item{disc_table}{A table with the following columns:
  feature: the column names of X.
  effect_size: an estimate of the difference between groups.
  t.stat: the t-statistic associated with the test on group differences.
  p.value: the p-value of the t-statistic.
  q.value: the multiple hypothesis corrected p-value.
  pert: the column names of Y
  }
}
}
\description{
Gets the biomarkers for each column in a matrix of responses using:
random_forest
lin_associations
discrete_test
}
