% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_achilles}
\alias{load_achilles}
\title{Loads omics features from taiga}
\usage{
load_achilles(gene)
}
\value{
a vector of gene cell line x gene effect for the given gene.
}
\description{
Loads omics features from taiga
}
