% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_omics}
\alias{load_omics}
\title{Loads omics features from taiga}
\usage{
load_omics()
}
\value{
a named list containing RNA expression, RPPA, Micro RNA, and Proteomics
}
\description{
Loads omics features from taiga
}
