% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{random_forest}
\alias{random_forest}
\title{Fits a random forest from a matrix or features X and a vector y.}
\usage{
random_forest(X, y, k = 10, n = 250)
}
\arguments{
\item{X}{n x m numerical matrix of features (missing values are allowed).}

\item{y}{length n vector of numerical values (missing values are allowed).}

\item{k}{integer number of cross validation cycles to perform.}

\item{n}{number of features to be considered in the model (after correlation filter).}
}
\value{
a list with components
\describe{
  \item{model_table}{A table with the following columns:
  feature: the column names of X.
  RF.imp.mean: an estimate of the importance of that feature for model accuracy.
  RF.imp.sd: the standard deviation of the importance estimate.
  RF.imp.stability: the proportion of models that used this feature.
  rank: the rank of the feature in terms of importance for the model.
  MSE: the mean-squared error of the model.
  MSE.se: the standard error of the MSE.
  R2: the \eqn{R^2} of the model.
  PearsonScore: the Pearson correlation of predicted and observed responses.
  }
  \item{predictions}{A vector of the responses predicted by the random forest.}
}
}
\description{
Fits a random forest from a matrix or features X and a vector y.
}
