% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_association.R
\name{discrete_test}
\alias{discrete_test}
\title{Performs t-tests on a matrix of binary features and a vector y.}
\usage{
discrete_test(X, y)
}
\arguments{
\item{X}{n x m matrix of binary features (missing values are allowed).}

\item{y}{length n vector of numerical values (missing values are allowed).}
}
\value{
A table with the following columns:
\describe{
feature: the column names of X. \cr
effect_size: an estimate of the difference between groups. \cr
t.stat: the t-statistic associated with the test on group differences. \cr
p.value: the p-value of the t-statistic. \cr
q.value: the multiple hypothesis corrected p-value. \cr
}
}
\description{
Performs t-tests on a matrix of binary features and a vector y.
}
