% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{get_biomarkers}
\alias{get_biomarkers}
\title{Gets the biomarkers for each column in a matrix of responses using:
random_forest
lin_associations
discrete_test}
\usage{
get_biomarkers(Y, p_cutoff = 0.1, out_path = NULL)
}
\arguments{
\item{Y}{n x p numerical matrix of responses to perturbations,
rownames must be Arxspan IDs (missing values are allowed).}

\item{p_cutoff}{maximum p-value in output tables (to keep them smaller), defaults to 0.1}

\item{out_path}{if specified, string path to folder to write results to as a .csv}
}
\value{
a list with components
\describe{
  \item{rf_table}{A table with the following columns: \cr
  feature: the column names of X. \cr
  RF.imp.mean: an estimate of the importance of that feature for model accuracy. \cr
  RF.imp.sd: the standard deviation of the importance estimate. \cr
  RF.imp.stability: the proportion of models that used this feature. \cr
  rank: the rank of the feature in terms of importance for the model. \cr
  MSE: the mean-squared error of the model. \cr
  MSE.se: the standard error of the MSE. \cr
  R2: the \eqn{R^2} of the model. \cr
  PearsonScore: the Pearson correlation of predicted and observed responses. \cr
  pert: the column names of Y \cr
  }
  \item{lin_table}{A table with the following columns: \cr
  feature:  column names of A \cr
  p.val: p-values computed for the null-hypothesis beta = 0 \cr
  q.val: q-values computed using Benjamini-Hotchberg method \cr
  n: number of non-missing values for each column of A \cr
  betahat: estimated linear coefficient controlled for W \cr
  sebetahat: standard error for the estimate of beta \cr
  sebetahat: standard error for the estimate of beta \cr
  lfdr: local and global fdr values computed using adaptive shrinkage as \cr
  qvalue: local and global fdr values computed using adaptive shrinkage as \cr
  PosteriorMean: moderated effect size estimates again computed with adaptive shrinkage \cr
  PosteriorSD: standard deviations of moderated effect size estimates \cr
  z.score: z-score PosteriorMean/PosteriorSD \cr
  }
  \item{disc_table}{A table with the following columns:
  feature: the column names of X. \cr
  effect_size: an estimate of the difference between groups. \cr
  t.stat: the t-statistic associated with the test on group differences. \cr
  p.value: the p-value of the t-statistic. \cr
  q.value: the multiple hypothesis corrected p-value. \cr
  pert: the column names of Y \cr
  }
}
}
\description{
Gets the biomarkers for each column in a matrix of responses using:
random_forest
lin_associations
discrete_test
}
