% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_association.R
\name{lin_associations}
\alias{lin_associations}
\title{Calculates the linear association between a matrix of features A and a vector y.}
\usage{
lin_associations(A, y, W = NULL, robust.se = F, scale.A = T,
  scale.y = F, is.dependent = T, var.th = 0, coef.var = F)
}
\arguments{
\item{A}{n x m numerical matrix of features (missing values are allowed).}

\item{y}{length n vector of numerical values (missing values are not allowed)}

\item{W}{n x p numerical matrix of confounders (missing values are not allowed).}

\item{robust.se}{boolean flag to control if homoskedastic or heteroskedastic.}

\item{scale.A}{boolean to control if the columns of A to be scaled or not.}

\item{scale.y}{boolen to control if the variance of y should be scaled to 1.}

\item{is.dependent}{should y be considered as a dependent or independent variable}

\item{var.th}{float minimum variation of included feature}

\item{coef.var}{boolean to control if the coefficient of variation is used.}
}
\value{
A table with the following columns:
\describe{
feature:  column names of A \cr
p.val: p-values computed for the null-hypothesis beta = 0 \cr
q.val: q-values computed using Benjamini-Hotchberg method \cr
n: number of non-missing values for each column of A \cr
betahat: estimated linear coefficient controlled for W \cr
sebetahat: standard error for the estimate of beta \cr
sebetahat: standard error for the estimate of beta \cr
lfdr: local and global fdr values computed using adaptive shrinkage as \cr
qvalue: local and global fdr values computed using adaptive shrinkage as \cr
PosteriorMean: moderated effect size estimates again computed with adaptive shrinkage \cr
PosteriorSD: standard deviations of moderated effect size estimates \cr
z.score: z-score PosteriorMean/PosteriorSD}
}
\description{
Calculates the linear association between a matrix of features A and a vector y.
}
